frontend(react): Modulated UploadButton into it's own component ui: Changed some text on the LandingPage and added a workable "Upload DICOM Files" button in the center of the table when there are no files uploaded, to be accompanied with the "No DICOM Data Available" message
