+++ frontend/src/components/workspace-components/table-utils/index.js
@@ -1,3 +1,4 @@
 export { default as handleDicomDataFetching } from './handleDicomDataFetching';
 export { default as handleDicomDelete } from './handleDicomDelete';
 export {formatName, formatDate, formatAge} from './formatHelpers'
+export { default as UploadButton } from './UploadButton';
